move subscriber to core and fix tests
