+++ packages/api/src/app.ts
@@ -1,7 +1,6 @@
 import 'module-alias/register';
-import { utils, config, database } from '@impactmarket/core';
+import { utils, config, database, subscriber } from '@impactmarket/core';
 import express from 'express';
-import job from './subscriber/job';
 
 import serverLoader from './server';
 
@@ -25,7 +24,7 @@ export async function startServer() {
     await serverLoader(app);
     utils.Logger.info('📡 Express server loaded');
 
-    job();
+    subscriber();
     utils.Logger.info('⏱️ Chain Subscriber starting');
 
     return app.listen(config.port, () => {
