+++ packages/core/tests/integration/subscriber/CommunityAdmin.json
@@ -1,0 +1,3 @@
+{
+    "bytecode": "0x608060405234801561001057600080fd5b5060405161080138038061080183398101604081905261002f91610054565b600080546001600160a01b0319166001600160a01b0392909216919091179055610084565b60006020828403121561006657600080fd5b81516001600160a01b038116811461007d57600080fd5b9392505050565b61076e806100936000396000f3fe608060405234801561001057600080fd5b506004361061004c5760003560e01c806322b5ed72146100515780638e2d90db146100805780639263454014610095578063a4937c51146100a8575b600080fd5b600054610064906001600160a01b031681565b6040516001600160a01b03909116815260200160405180910390f35b61009361008e3660046102ff565b6100bb565b005b6100936100a33660046102dd565b610143565b6100936100b636600461034d565b61017a565b6000826040516100ca9061020c565b6100d49190610408565b604051809103906000f0801580156100f0573d6000803e3d6000fd5b509050816001600160a01b0316816001600160a01b03167f1c73dbd6259eae88a71a197a742361367d2811051b3de2b22953065b876247af856040516101369190610408565b60405180910390a3505050565b6040516001600160a01b038216907fa285b77d62d36ec6881b2cbc019c53874eb061798176f3577f2beab0848d16c690600090a250565b6000886040516101899061020c565b6101939190610408565b604051809103906000f0801580156101af573d6000803e3d6000fd5b509050806001600160a01b03167fd975f7d7350dc54e14ea0400bad2e29483d76cb8810f36f1f520d5c44c1dcf028a8a8a8a8a8a8a8a6040516101f998979695949392919061041b565b60405180910390a2505050505050505050565b6102bc8061047d83390190565b80356001600160a01b038116811461023057600080fd5b919050565b600082601f83011261024657600080fd5b8135602067ffffffffffffffff8083111561026357610263610466565b8260051b604051601f19603f8301168101818110848211171561028857610288610466565b604052848152838101925086840182880185018910156102a757600080fd5b600092505b858310156102d1576102bd81610219565b8452928401926001929092019184016102ac565b50979650505050505050565b6000602082840312156102ef57600080fd5b6102f882610219565b9392505050565b6000806040838503121561031257600080fd5b823567ffffffffffffffff81111561032957600080fd5b61033585828601610235565b92505061034460208401610219565b90509250929050565b600080600080600080600080610100898b03121561036a57600080fd5b883567ffffffffffffffff81111561038157600080fd5b61038d8b828c01610235565b9b60208b01359b5060408b01359a60608101359a506080810135995060a0810135985060c0810135975060e0013595509350505050565b600081518084526020808501945080840160005b838110156103fd5781516001600160a01b0316875295820195908201906001016103d8565b509495945050505050565b6020815260006102f860208301846103c4565b600061010080835261042f8184018c6103c4565b602084019a909a52505060408101969096526060860194909452608085019290925260a084015260c083015260e090910152919050565b634e487b7160e01b600052604160045260246000fdfe608060405234801561001057600080fd5b506040516102bc3803806102bc83398101604081905261002f916100bf565b60005b815181101561009c57600082828151811061004f5761004f6101b3565b60209081029190910181015182546001810184556000938452919092200180546001600160a01b0319166001600160a01b03909216919091179055806100948161018a565b915050610032565b50506101df565b80516001600160a01b03811681146100ba57600080fd5b919050565b600060208083850312156100d257600080fd5b82516001600160401b03808211156100e957600080fd5b818501915085601f8301126100fd57600080fd5b81518181111561010f5761010f6101c9565b8060051b604051601f19603f83011681018181108582111715610134576101346101c9565b604052828152858101935084860182860187018a101561015357600080fd5b600095505b8386101561017d57610169816100a3565b855260019590950194938601938601610158565b5098975050505050505050565b60006000198214156101ac57634e487b7160e01b600052601160045260246000fd5b5060010190565b634e487b7160e01b600052603260045260246000fd5b634e487b7160e01b600052604160045260246000fd5b60cf806101ed6000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063bcd674a214602d575b600080fd5b603c60383660046081565b6058565b6040516001600160a01b03909116815260200160405180910390f35b60008181548110606757600080fd5b6000918252602090912001546001600160a01b0316905081565b600060208284031215609257600080fd5b503591905056fea2646970667358221220f72a5fd07e9fc150122862e824750cd5b0d033692269fbfd76709b555b3a7b4064736f6c63430008050033a26469706673582212209712972826a92b16e681bb014b3b0827ecf58c75e4bd78b4cf194e16dd21a37164736f6c63430008050033"
+}
